pyproject.toml
[project]
name = "tiktok-analysis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = []

[dependency-groups]
dev = [
    "chromadb>=1.3.0",
    "ipykernel>=7.1.0",
    "langchain>=1.0.3",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.0.1",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "selenium>=4.38.0",
    "seleniumbase>=4.44.4",
]
